Update Pkgfile
The english version 0.51-1 is for the aspell version 0.50 and below and is incompatible and doesn't work with aspell version 0.60 and higher. I've updated the package file with correct dictionary version for aspell 0.60 and up. I'm not sure about the rest of the languages. For clarification, you can consult this page,
https://ftp.gnu.org/gnu/aspell/dict/0index.html